+++ store/src/test/java/com/zextras/mailbox/util/MailboxSetupHelper.java
@@ -14,7 +14,6 @@ import com.zextras.mailbox.util.InMemoryLdapServer.Builder;
 import com.zimbra.common.account.ZAttrProvisioning;
 import com.zimbra.common.calendar.WellKnownTimeZones;
 import com.zimbra.common.localconfig.LC;
-import com.zimbra.cs.account.LdapProvisioningWithMockMime;
 import com.zimbra.cs.account.Provisioning;
 import com.zimbra.cs.account.accesscontrol.RightManager;
 import com.zimbra.cs.db.DbPool;
